DomStorageContext: Fix a crash when scavenging session storages.
The scavenging will go wrong at least if StartScavengingUnusedSessionStorage is
called more than once. A separate CL will add debugging information to find out
if that happens and why.

Kudos to eroman@chromium.org for finding the cause.

BUG=145068


Review URL: https://chromiumcodereview.appspot.com/10917034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154572 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/dom_storage/dom_storage_context.cc
@@ -305,7 +305,8 @@ void DomStorageContext::StartScavengingUnusedSessionStorage() {
 
 void DomStorageContext::FindUnusedNamespaces() {
   DCHECK(session_storage_database_.get());
-  DCHECK(!scavenging_started_);
+  if (scavenging_started_)
+    return;
   scavenging_started_ = true;
   std::set<std::string> namespace_ids_in_use;
   for (StorageNamespaceMap::const_iterator it = namespaces_.begin();
@@ -356,6 +357,8 @@ void DomStorageContext::DeleteNextUnusedNamespace() {
 }
 
 void DomStorageContext::DeleteNextUnusedNamespaceInCommitSequence() {
+  if (deletable_persistent_namespace_ids_.empty())
+    return;
   const std::string& persistent_id = deletable_persistent_namespace_ids_.back();
   session_storage_database_->DeleteNamespace(persistent_id);
   deletable_persistent_namespace_ids_.pop_back();
